src/backend/requirements.txt
aiohttp==3.9.1
aiosignal==1.3.2
annotated-types==0.7.0
anyio==3.7.1
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
distro==1.9.0
fastapi==0.104.1
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.9.0
multidict==6.1.0
numpy==2.2.4
openai==1.68.2
pandas==2.2.3
prometheus-client==0.19.0
propcache==0.3.0
psutil==5.9.6
pydantic==2.10.6
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.1
requests==2.31.0
six==1.17.0
sniffio==1.3.1
starlette==0.27.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.3.0
uvicorn==0.23.2
websockets==11.0.3
yarl==1.18.3
